backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==2.0.12
Django==4.2
djangorestframework==3.15.2
django-cors-headers==3.13.0
django-filter==23.1
djoser==2.1.0
drf-extra-fields
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==0.13.1
psycopg2-binary==2.9.3
pillow==11.0.0
py==1.11.0
PyJWT==2.7.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2024.2
python-dotenv==1.0.1
requests==2.26.0
requests-oauthlib==2.0.0
sqlparse==0.5.1
toml==0.10.2
typing_extensions==4.12.2
urllib3==1.26.20
